[Persistence][15] Extract reusable CoreData helpers into a separate file
